+++ sys/sys_local.h
@@ -61,9 +61,11 @@ public:
 	virtual bool			LockMemory( void *ptr, int bytes );
 	virtual bool			UnlockMemory( void *ptr, int bytes );
 
-	virtual int				DLL_Load( const char *dllName );
-	virtual void *			DLL_GetProcAddress( int dllHandle, const char *procName );
-	virtual void			DLL_Unload( int dllHandle );
+	// Techyon RB: 64 bit fixes, changed int to intptr_t
+	virtual intptr_t		DLL_Load( const char *dllName );
+	virtual void *			DLL_GetProcAddress( intptr_t dllHandle, const char *procName );
+	virtual void			DLL_Unload( intptr_t dllHandle );
+	// Techyon END
 	virtual void			DLL_GetFileName( const char *baseName, char *dllName, int maxLength );
 
 	virtual sysEvent_t		GenerateMouseButtonEvent( int button, bool down );
